+++ cli/src/commands/router/types/types.ts
@@ -1,0 +1,6 @@
+export interface FetchRouterConfigResult {
+  splitConfigLoading: boolean;
+  routerConfig: string;
+  featureFlags?: Map<string, string>;
+  mapper?: Record<string, string>;
+}